Add git version description as docker images tags

+++ Makefile
@@ -1,5 +1,7 @@
 RACKATTACK_PHYSICAL_DOCKER_CIDFILE = /var/lib/rackattackphysical/cid
 UPSETO_REQUIREMENTS_FULFILLED = $(shell upseto checkRequirements 2> /dev/null; echo $$?)
+VERSION=$(shell git describe --tags --dirty)
+
 all: validate_requirements unittest build check_convention
 
 clean:
@@ -101,7 +103,8 @@ build/rackattack-physical-reclamation.dockerfile: docker/rackattack-physical-rec
 
 .PHONY: rackattack-physical-docker-image
 rackattack-physical-docker-image: build/rackattack-physical.dockerfile
-	@echo "Building the rackattack-physical docker image..."
+ifeq ($(shell sudo docker images | egrep -Ec "^rackattack-physical[ ]+$(VERSION)" | xargs echo -n),0)
+	$(info Building the rackattack-physical docker image for version '$(VERSION)'...)
 	-rm -rf build/rackattack-{physical,virtual,api}
 	-mkdir build
 	cp -rf {../,build/}rackattack-virtual
@@ -110,12 +113,19 @@ rackattack-physical-docker-image: build/rackattack-physical.dockerfile
 	cp -rf . /tmp/rackattack-physical.temp
 	cp -rf /tmp/rackattack-physical.temp build/rackattack-physical
 	rm -rf /tmp/rackattack-physical.temp
-	@docker build -f $< -t rackattack-physical build
+	@docker build -f $< -t "rackattack-physical:$(VERSION)" build
+else
+	$(info It seems that a rackattack-physical Docker image for version '$(VERSION)' already exists. Skipping build.)
+endif
 
 .PHONY: rackattack-physical-reclamation-docker-image
 rackattack-physical-reclamation-docker-image: build/rackattack-physical-reclamation.dockerfile
-	@echo "Building the rackattack-physical-reclamation docker image..."
-	@docker build -f $< -t rackattack-physical-reclamation build
+ifeq ($(shell sudo docker images | egrep -Ec "^rackattack-physical-reclamation[ ]+$(VERSION)" | xargs echo -n),0)
+	$(info Building the rackattack-physical-reclamation docker image for version '$(VERSION)'...)
+	@docker build -f $< -t "rackattack-physical-reclamation:$(VERSION)" build
+else
+	$(info It seems that a rackattack-physical-reclamation Docker image for version '$(VERSION)' already exists. Skipping build.)
+endif
 
 build/pipework:
 	-wget --no-check-certificate https://raw.github.com/jpetazzo/pipework/master/pipework -O build/pipework
@@ -123,12 +133,12 @@ build/pipework:
 
 .PHONY: run-rackattack-physical-docker-container
 run-rackattack-physical-docker-container: rackattack-physical-docker-image build/pipework
-ifneq ($(shell docker ps | grep -c "rackattack-physical:" | xargs echo -n),0)
+ifneq ($(shell docker ps | grep -c "rackattack-physical:$(VERSION)" | xargs echo -n),0)
 	$(error Cannot start rackattack while another rackattack container is running.)
 	exit 1
 endif
 	-rm "$(RACKATTACK_PHYSICAL_DOCKER_CIDFILE)"
-	docker run -d=true -v /etc/rackattack-physical:/etc/rackattack-physical -v /usr/share/rackattack.physical/reclamation_requests_fifo:/usr/share/rackattack.physical/reclamation_requests_fifo -v /usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo:/usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo -v /var/lib/rackattackphysical/:/var/lib/rackattackphysical/ -p 1013:1013 -p 1014:1014 -p 1015:1015 -p 1016:1016 -p 67:67/udp -p 69:69 -p 53:53/udp --cap-add NET_ADMIN --cidfile="$(RACKATTACK_PHYSICAL_DOCKER_CIDFILE)" rackattack-physical
+	docker run -d=true -v /etc/rackattack-physical:/etc/rackattack-physical -v /usr/share/rackattack.physical/reclamation_requests_fifo:/usr/share/rackattack.physical/reclamation_requests_fifo -v /usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo:/usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo -v /var/lib/rackattackphysical/:/var/lib/rackattackphysical/ -p 1013:1013 -p 1014:1014 -p 1015:1015 -p 1016:1016 -p 67:67/udp -p 69:69 -p 53:53/udp --cap-add NET_ADMIN --cidfile="$(RACKATTACK_PHYSICAL_DOCKER_CIDFILE)" "rackattack-physical:$(VERSION)"
 	@echo "Setting up networking for the rackattack-physical container..."
 	@UPSETO_JOIN_PYTHON_NAMESPACES=Yes PYTHONPATH=. python rackattack/physical/setup_networking_for_docker_idempotently.py "`cat /var/lib/rackattackphysical/cid`" build/pipework
 	@echo "Done."
@@ -139,7 +149,7 @@ ifneq ($(shell docker ps | grep -c "rackattack-physical-reclamation:"),0)
 	$(error Cannot start rackattack while another rackattack container is running.)
 	exit 1
 endif
-	docker run -d=true -v /etc/rackattack-physical:/etc/rackattack-physical -v /usr/share/rackattack.physical/reclamation_requests_fifo:/usr/share/rackattack.physical/reclamation_requests_fifo -v /usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo:/usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo rackattack-physical-reclamation
+	docker run -d=true -v /etc/rackattack-physical:/etc/rackattack-physical -v /usr/share/rackattack.physical/reclamation_requests_fifo:/usr/share/rackattack.physical/reclamation_requests_fifo -v /usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo:/usr/share/rackattack.physical/soft_reclamations_failure_msg_fifo "rackattack-physical-reclamation:$(VERSION)"
 
 .PHONY: install_as_docker_containers
 install_as_docker_containers: run-rackattack-physical-reclamation-container run-rackattack-physical-reclamation-container
